refactor: move tabs util func to a separate file
