+++ src/utils/index.ts
@@ -9,3 +9,4 @@ export {
 } from "./trackTime";
 export { validateSpaceTitle } from "./validate";
 export { cn } from "./cn";
+export { getFaviconUrl, sortTabs } from "./tabs";
